PythonAI/requirements.txt
langchain-community
langchain-core
langchain-openai
python-dotenv
pypdf
pydantic>=2.0
fastapi
uvicorn[standard]
python-multipart
redis

